requirements.txt
# Requires Python >=3.6.8
cycler==0.11.0
dataclasses==0.8
joblib==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.17.0
sklearn==0.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.1.1
